backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dateutil==2.8.2
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
PyJWT==2.8.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2

# Testing Dependencies
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.2.2

# Code Quality
flake8==6.0.0
black==23.7.0
isort==5.12.0
mypy==1.5.1

# Security Dependencies
flask-limiter==3.5.0
flask-talisman==1.1.0
safety==2.3.5
bandit==1.7.5

# Monitoring Dependencies
psutil==5.9.6

# Production Dependencies  
gunicorn==21.2.0

# Development Dependencies
python-dotenv==1.0.0
